examples/requirements.txt
psycopg2-binary
numpy
requests